Small Window tweaks to fit content
My monitor apparently isn't as big as the authors... I had to keep on resizing the window just to close out of the program. I think there's probably a better way to choose the default size, but I think this is an improvement for now.  Also, some of the text in the plot fit options went over the bounds of the window, and the window didn't have a resizer. Setting the window size to -1 allowed for Auto-sizing appropriate for the content.